[CARE-46] refactor : set access path #9
